+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `laravel-artisan-runner` will be documented in this file.
 
-## Unreleased
+## 1.2.2 - 2026-06-12
 
 ### Fixed